Move all api tests related to contacts routes to /routers/contacts package
