+++ tests/unit/routers/contacts/test_api_delete.py
