+++ tests/unit/routers/contacts/test_api_patch.py
