+++ tests/unit/routers/contacts/test_bulk.py
